fix: run touchscreen test in a subprocess to isolate GDK crashes
The touchscreen test reliably SIGSEGVs in GTK/GDK touch event delivery
after a USB pointer device is unplugged, on the first touch tap of a
target. Multiple in-process workarounds (deferred fullscreen, deduped
CSS provider, display.sync(), Gtk.Box+explicit GestureClick, single
gesture on the container with manual hit-testing) all failed; the
crash is somewhere in GDK's touch routing that we cannot reach from
Python.

Extract TouchscreenTest into a standalone runner script and invoke it
via subprocess from on_touchscreen_clicked. The runner writes "pass"
or "fail" to stdout on completion; the parent reads it via
subprocess.run on a background thread and updates the checkbox via
GLib.idle_add. A SIGSEGV in the subprocess now fails the test instead
of killing the provisioning app.
